Added phase1_PixelEvents.R
This is an additional piece of R-code for Phase1. It converts a list of dead pixel coordinates into an object that contains both this and a list of coordinates of pixels representing damage events rather than individual pixels. Damage events can be isolated dead pixels, but in the event of larger clusters or lines of dead pixels, events are used to represent them in the form of just one dead pixel (which is constructed in the code depending on the type of the cluster or line).
Output directories need to be specified before running it.