Add files via upload
This is the new version of the code that modifies the area to be analysed. 
It still has variables and style as in original code; you're welcome to adjust this for consistency.
I didn't set any path for the plots. Please add this consistent with what has been done for other plot functions in the package. 

Apart from two plots, the output of the function is a new point pattern like pePPP (with slots for Pixels and Events) but the high area densities cut out. 

Ideally, for later use, it would also output a binary masks of the detector area that has 0 for the areas cut out (convex hulls of high damage areas) and 1 elsewhere. If you know of an easy way to do that, please add that; otherwise I'll get back to that at a later stage.